+++ windows/ble_peripheral_plugin.cpp
@@ -1,24 +1,15 @@
 #include "ble_peripheral_plugin.h"
-// This must be included before many other Windows headers.
 #include <windows.h>
-#include <winrt/Windows.Foundation.h>
-#include <winrt/Windows.Foundation.Collections.h>
-#include <winrt/Windows.Storage.Streams.h>
-#include <winrt/Windows.Devices.Radios.h>
-#include <winrt/Windows.Devices.Bluetooth.h>
-#include <winrt/Windows.Devices.Bluetooth.Advertisement.h>
-#include <winrt/Windows.Devices.Bluetooth.GenericAttributeProfile.h>
-#include <winrt/Windows.Devices.Enumeration.h>
-
-// For getPlatformVersion; remove unless needed for your plugin implementation.
-#include <VersionHelpers.h>
-
-#include <flutter/method_channel.h>
 #include <flutter/plugin_registrar_windows.h>
-#include <flutter/standard_method_codec.h>
-
+#include <map>
 #include <memory>
 #include <sstream>
+#include <algorithm>
+#include <iomanip>
+#include <thread>
+#include <regex>
+#include "Utils.h"
+
 #include "BlePeripheral.g.h"
 
 namespace ble_peripheral
@@ -26,7 +17,6 @@ namespace ble_peripheral
   using ble_peripheral::BleCallback;
   using ble_peripheral::BlePeripheralChannel;
   using ble_peripheral::ErrorOr;
-
   std::unique_ptr<BleCallback> bleCallback;
 
   // static
@@ -38,7 +28,7 @@ namespace ble_peripheral
     registrar->AddPlugin(std::move(plugin));
   }
 
-  BlePeripheralPlugin::BlePeripheralPlugin(){}
+  BlePeripheralPlugin::BlePeripheralPlugin() {}
 
   BlePeripheralPlugin::~BlePeripheralPlugin() {}
 
@@ -47,68 +37,61 @@ namespace ble_peripheral
     auto bluetoothAdapter = co_await BluetoothAdapter::GetDefaultAsync();
     bluetoothRadio = co_await bluetoothAdapter.GetRadioAsync();
     bluetoothLEPublisher = BluetoothLEAdvertisementPublisher();
-    // status_changed_token = bluetoothLEPublisher.StatusChanged(
-    //     winrt::auto_revoke,
-    //     [this](BluetoothLEAdvertisementPublisher sender, BluetoothLEAdvertisementPublisherStatusChangedEventArgs args)
-    //     {
-    //       // Handle status changed event
-    //       // std::cout << args.Status() << std::endl;
-    //       std::cout << "Advertisement publisher status changed:" << std::endl;
-    //     });
-    // TO send event to flutter
-    // bleCallback->OnBleStateChange(
-    //     true,
-    //     [this]()
-    //     {
-    //       // on_success callback
-    //       std::cout << "Advertising started successfully" << std::endl;
-    //     },
-    //     [this](const FlutterError &error)
-    //     {
-    //       // on_error callback
-    //       std::cerr << "Error starting advertising: " << error.message() << std::endl;
-    //     });
+    bluetoothLEPublisher.StatusChanged([this](auto &&, auto &&)
+                                       {
+                                         std::cout << "AdvertisingStatusChanged" << std::endl;
+                                         //  auto status = bluetoothLEPublisher.Status();
+                                         //  bleCallback->OnAdvertisingStarted(nullptr, SuccessCallback, ErrorCallback);
+                                       });
   }
 
   std::optional<FlutterError> BlePeripheralPlugin::Initialize()
   {
     InitializeAdapter();
-    std::cout << "Initialize called" << std::endl;
-
     return std::nullopt;
   }
 
-  ErrorOr<bool> BlePeripheralPlugin::IsAdvertising()
+  ErrorOr<std::optional<bool>> BlePeripheralPlugin::IsAdvertising()
   {
-    std::cout << "IsAdvertising called" << std::endl;
-    return true;
+    auto status = bluetoothLEPublisher.Status();
+    return ErrorOr<std::optional<bool>>(status == BluetoothLEAdvertisementPublisherStatus::Started);
   }
 
   ErrorOr<bool> BlePeripheralPlugin::IsSupported()
   {
-    std::cout << "IsSupported called" << std::endl;
-    return true;
+    return bluetoothLEPublisher != nullptr;
   }
 
-  std::optional<FlutterError> BlePeripheralPlugin::AddServices(const flutter::EncodableList &services)
+  ErrorOr<bool> BlePeripheralPlugin::AskBlePermission()
   {
-    std::cout << "AddServices called" << std::endl;
+    std::cout << "AskBlePermission called" << std::endl;
+    return true;
+  };
+
+  std::optional<FlutterError> BlePeripheralPlugin::AddService(const BleService &service)
+  {
+    auto serviceUuid = service.uuid();
+    std::cout << "Adding service " << serviceUuid << std::endl;
+    bluetoothLEPublisher.Advertisement().ServiceUuids().Append(uuid_to_guid(serviceUuid));
+    bleCallback->OnServiceAdded(serviceUuid, nullptr, SuccessCallback, ErrorCallback);
     return std::nullopt;
-  }
+  };
 
   std::optional<FlutterError> BlePeripheralPlugin::StartAdvertising(
       const flutter::EncodableList &services,
-      const std::string &local_name)
+      const std::string &local_name,
+      const int64_t *timeout,
+      const ManufacturerData *manufacturer_data,
+      bool add_manufacturer_data_in_scan_response)
   {
+    // Advertisement::BluetoothLEManufacturerData manufacturerData = Advertisement::BluetoothLEManufacturerData();
+    // manufacturerData.CompanyId(0xFFFE);
+    // advertisement.ManufacturerData().Append(manufacturerData);
+    BluetoothLEAdvertisement advertisement = bluetoothLEPublisher.Advertisement();
+    advertisement.Flags(BluetoothLEAdvertisementFlags::GeneralDiscoverableMode);
 
-    Advertisement::BluetoothLEManufacturerData manufacturerData = Advertisement::BluetoothLEManufacturerData();
-    manufacturerData.CompanyId(0xFFFE);
-    auto dataWriter = DataWriter();
-    dataWriter.WriteBytes("Test");
-    manufacturerData.Data(dataWriter.DetachBuffer());
-
-    bluetoothLEPublisher.Advertisement().ManufacturerData().Append(manufacturerData);
     bluetoothLEPublisher.Start();
+    advertising = true;
     std::cout << "StartAdvertising called" << std::endl;
     return std::nullopt;
   }
@@ -116,14 +99,15 @@ namespace ble_peripheral
   std::optional<FlutterError> BlePeripheralPlugin::StopAdvertising()
   {
     std::cout << "StopAdvertising called" << std::endl;
-    bluetoothLEPublisher.Advertisement().ManufacturerData().Clear();
+    // bluetoothLEPublisher.Advertisement().ManufacturerData().Clear();
     bluetoothLEPublisher.Stop();
+    advertising = false;
     return std::nullopt;
   }
 
   std::optional<FlutterError> BlePeripheralPlugin::UpdateCharacteristic(
-      const BleCentral &central,
-      const BleCharacteristic &characteristic,
+      const std::string &devoice_i_d,
+      const std::string &characteristic_id,
       const std::vector<uint8_t> &value)
   {
     std::cout << "UpdateCharacteristic called" << std::endl;