+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0.0
+
+- Refactored Apis
+- Fix android permission issues
+- Update callback handlers issue
+- Add onMtuChange callback
+- Fix AddServices
+- Update Readme
+- Some more bug fixes
+
 ## 0.0.1
 
-* Initial version
+- Initial version